requirements.txt
requests>=2.28
jsonschema>=4.0
lxml>=4.9
PyYAML>=6.0
